Added basic implementation of the IScoreParser interface
Currently the implementation makes a _lot_ of assumptions about the format of the input string. It also will not cope with being able to use something like e.g. playing cards, dice other than six-sided and so on. For now we're just proving all the moving parts will fit together neatly.