Implemented a type to select a category by name in a case-insensitive fashion
This is intended to extract the predicate for `ICategory` selection from the `GameScorer`. That makes the `GameScorer` easier to test, and gives us more flexibility with our game rules e.g. the game might randomly choose a category, or choose the category that would have the highest score for that round etc.