Started to make ScoreParser less brittle by introducing a very basic regex
For now this doesn't get us anything we didn't already have i.e. we can only find single digits, and we expect _exactly_ five that _must_ be separated by ", " and the entire substring delimited by parentheses with no extraneous whitespace. However, we're now in a position where we can start building this out to account for more variation in the string.

The significant improvement this gives us over what we had before is the removal of many of the magic numbers indexing into the string.