Made test data for GetScoreValues strongly typed
By introducing a class to represent a singe test case we can then construct instances of that type and its constructor ensures we have the correct parameters in the correct order, unlike an `object[]`. We also get the advantage of the parameters being named so we can get a better idea of what we're actually passing!

Because the test case implements an implicit conversion to `object[]` we can still return it in the method that builds all of the test data 😃