Made parsing of category name from a score string more robust
We no longer assume the score is going to start at a fixed point in the string. We only assume it is going to come after the values end i.e. when we find the close parenthesis. This is important now the `ScoreParser` can cope with different numbers of dice, and with dice that have ten or more sides.