Simplified ScoreParserTests by replacing unnecessary use of MemberData with InlineData
For testing parsing of the values we _had_ to use `MemberData` because we needed an `IEnumerable<int>` as our expected , and we cannot pass an array to an attribute as it is not a constant value. We don't have this problem when testing the category name. The expected value is simply a `string` which we _can_ pass to an attribute. Therefore we can use `InlineData` and remove the type we were using to give us strong typing of the test parameters.