Implemented TwoPairsCategory
The specification was a little unclear about how to handle certain cases such as when we have a round that would technically match a Yahtzee or Full House. How I have chosen to handle these has been explicitly defined in the unit tests. I have also made the assumption this specific category is to be used for "classic" Yahtzee i.e. with rounds of five dice. As such cases such as six dice with two or three pairs do not score. Again, this is defined in the unit tests.