Made regular expression used to parse score values shorter and simpler
By switching to using the `Split` method, rather than the `Matches` method we can simplify the regular expression. We no longer need the look ahead, or look behind, and we don't need the named group either. We can switch to using a simple alternation of the start of the score string (e.g. opening paren), the delimiter (i.e. a comma and optionally a space), or the end (the close paren). By including all other characters when we're looking for the closing paren we don't have to remove the category from the results before we parse integers.

We _have_ lost a couple of things here. We're no longer guaranteed to capture digits so `int.Parse` has more potential to fail. We're also no longer guaranteed to start with an opening paren, and end with a closing paren.